refactor search imp and xml. refactor data local and remote. fix bugs.

+++ presentation/src/main/java/com/cheesecake/presentation/screens/country/countryLeagues/CountryLeaguesViewModel.kt
@@ -26,7 +26,7 @@ class CountryLeaguesViewModel @Inject constructor(
     private fun onSuccess(leagues: List<League>) {
         _state.update { countryLeaguesUIState ->
             countryLeaguesUIState.copy(
-                leagues = leagues.map { it.toLeagueUIState { onLeagueClick(it.leagueId, it.season.toInt()) } },
+                leagues = leagues.map { it.toLeagueUIState { onLeagueClick(it.leagueId, it.season) } },
                 isLoading = false
             )
         }
